requirements.txt
Django==5.1.3
django-jazzmin==3.0.1
django-redis==5.4.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
ipython==8.29.0
pillow==11.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
